==> 08 - Kaggle - bike share system - Adding customer-value-logarithm average to the features.ipynb <==
# 08 - Kaggle - bike share system - Adding customer-value-logarithm average to the features


For problem formulation refer to **"01 - Kaggle - bike share system - problem formulation.ipynb"**.
In section **"02 - Kaggle - bike share system - Data preprocessing.ipynb"** we transformed the raw data and extracted time, date, and dummy matrices. The results are stored in two formats:
 * In `train_prep_orig.csv` and `test_prep_orig.csv` the categorical data are in the original form.
 * In `train_prep_dum.csv` and `test_prep_dum.csv` the categorical data are converted to dummy matrices. 

In section **03 - Kaggle - bike share system - data visualization.ipynb** we ploted the average of customers at different time periods over 2011 and 2012 and discused the pattern of customer behavior. We concluded by a decision to consider these average values as new features of the problem so that the machine learning model will be able to use them as the basis values and the other features apply the necessary correction to make use closer to the actual values. That is the job of this section.

We split the hours time into 6 chunks of 4 hours periods. Here was the observation of customer behavior:
* 1: **[2:00 am, 3:00 am, 4:00 am, 5:00 am]**  ------> both casual and registered (and therefore total) customers are highest during the weekend.
* 2: ** [6:00 am, 7:00 am, 8:00 am, 9:00 am]**  ------> Casual customers behave for all day alsmot within the same range below the week average. Registered customers use the system on average more during the workdays above the the week average and the usage during the weekends in below the week average.
* 3: ** [10:00 am, 11:00 am, 12:00 pm, 1:00 pm]**  --> both casual and registered (and therefore total) customers are highest during the weekend.
* 4: ** [2:00 pm, 3:00 pm, 4:00 pm, 5:00 pm] **  ------> All are above the week average. The casual and total is highest during the weekend. For registered, all days are of the same order. 
* 5: ** [6:00 pm, 7:00 pm, 8:00 pm, 9:00 pm] **  ------> no comment.
* 6: ** [10:00 pm, 11:00 pm, 12:00 am, 1:00 pm] ** --> All are below the week average and the average number of customers is highest during the weekend.

Since we do not have the number of customers from day 20 to the end of each month, we have about three isntances of a weekday for each month and year. We take the average number of customers during these days and use it as a basis. We call the new features **avg_casual**, **avg_registered** and **avg_tot**. We finally, update the data sets as:
 * In `train_prep_orig_avg.csv` and `test_prep_orig_avg.csv` the categorical data are in the original form.
 * In `train_prep_dum_avg.csv` and `test_prep_dum_avg.csv` the categorical data are converted to dummy matrices. 



### Basic settings and importing the libraries

# Resets the namespace by removing all names defined by the user without asking for confirmation
%reset -f


# Panas is used for DataFrame
import pandas as pd

# NumPy is used for manipulating arrays
import numpy as np

# MatPlotLib is used for plotting
import matplotlib.pyplot as plt
import matplotlib.patches as mpatches

# the output of plotting commands is displayed inline directly below the code cell that produced it.
%matplotlib inline

# Seaborn is used for statistical plotting
import seaborn as sns

# Used for display dataframes as html tables
from IPython.display import display

### Importing the train data from `train_prep_orig.csv`

#Load train data
data_train = pd.read_csv('data/train_prep_orig.csv')

print "The shape of the train dataset:", data_train.shape
display(data_train.head())


## Adding the log1p values of target variables

data_train['l_casual']=np.log1p(data_train['casual'])
data_train['l_registered']=np.log1p(data_train['registered'])
data_train['l_tot']=np.log1p(data_train['tot'])

display(data_train.head())


## features and dictionaries

cat_var = ['year','season', 'month', 'weekday', 'hour', 'workingday', 'holiday', 'weather']
num_var = ['temp', 'atemp', 'humidity', 'windspeed']
target_var = ['casual', 'registered', 'tot']
target_l_var = ['l_casual', 'l_registered', 'l_tot']

weekday_dic = {0:'Monday',
              1:'Tuesday',
              2:'Wednesday',
              3:'Thursday',
              4:'Friday',
              5:'Saturday',
              6:'Sunday'}

### Grouping by ['`year`','`month`','`weekday`','`period`']

Now, we would like to learn how does it changes acros different months, weekdays and day periods. We only work with `mean` here .

hours = np.array(data_train.hour)

data_train['periods']=np.where( (2 <= hours) &  (hours <= 5), 1,
                        np.where( (6 <= hours) &  (hours <= 9), 2,
                                 np.where( (10 <= hours) &  (hours <= 13), 3,
                                          np.where( (14 <= hours) &  (hours <= 17), 4,
                                                  np.where( (18 <= hours) &  (hours <= 23), 5, 6)
                                                  )
                                         )
                                )
                        )
display(data_train.head(20))

groupby_year_month_weekday_periods = data_train.groupby(['year','month','weekday','periods']).mean()[target_l_var]
display(groupby_year_month_weekday_periods.head(40))

data_train['avg_l_casual'] = data_train['avg_l_registered'] = data_train['avg_l_tot'] = np.zeros(data_train.shape[0])
print data_train.shape[0]
data_train.head()

for i in range(data_train.shape[0]):
    if i % 1000 == 0: print i
    data_train.loc[i,'avg_l_casual'] = groupby_year_month_weekday_periods.xs(int(data_train.loc[i].year), level ='year'
                                     ).xs(int(data_train.loc[i].month), level ='month'
                                         ).xs(int(data_train.loc[i].weekday), level ='weekday'
                                             ).loc[int(data_train.loc[i].periods)]['l_casual']

for i in range(data_train.shape[0]):
    if i % 1000 == 0: print i
    data_train.loc[i,'avg_l_registered'] = groupby_year_month_weekday_periods.xs(int(data_train.loc[i].year), 
                                                                               level ='year'
                                     ).xs(int(data_train.loc[i].month), level ='month'
                                         ).xs(int(data_train.loc[i].weekday), level ='weekday'
                                             ).loc[int(data_train.loc[i].periods)].l_registered


for i in range(data_train.shape[0]):
    if i % 1000 == 0: print i
    data_train.loc[i,'avg_l_tot'] = groupby_year_month_weekday_periods.xs(int(data_train.loc[i].year), level ='year'
                                     ).xs(int(data_train.loc[i].month), level ='month'
                                         ).xs(int(data_train.loc[i].weekday), level ='weekday'
                                             ).loc[int(data_train.loc[i].periods)].l_tot

data_train.loc[130:140]


data_train=data_train.astype('float')

data_train.dtypes

### Saving `train_prep_orig_l_avg.csv`

cat_var = ['year','season', 'month', 'weekday', 'hour', 'workingday', 'holiday', 'weather']
num_var = ['temp', 'atemp', 'humidity', 'windspeed']
avg_var = ['avg_casual','avg_registered','avg_tot']
avg_l_var = ['avg_l_casual','avg_l_registered','avg_l_tot']
target_var = ['casual', 'registered', 'tot']
target_l_var = ['l_casual', 'l_registered', 'l_tot']



display(data_train[num_var+cat_var+avg_l_var+target_l_var].head())

data_train[num_var+cat_var+avg_l_var+target_l_var].to_csv('data/train_prep_orig_l_avg.csv', index=False)

The next step is to add the new features to the test data set. First we load the data set and group by ['`year`','`month`','`weekday`','`period`'].

data_test_avg = pd.read_csv('data/test_prep_orig.csv')
data_test_avg.head()

hours = np.array(data_test_avg.hour)

data_test_avg['periods']=np.where( (2 <= hours) &  (hours <= 5), 1,
                           np.where( (6 <= hours) &  (hours <= 9), 2,
                                 np.where( (10 <= hours) &  (hours <= 13), 3,
                                          np.where( (14 <= hours) &  (hours <= 17), 4,
                                                  np.where( (18 <= hours) &  (hours <= 23), 5, 6)
                                                  )
                                         )
                                )
                        )
display(data_test_avg.head(10))

data_test_avg['avg_l_casual'] = data_test_avg['avg_l_registered'] = data_test_avg['avg_l_tot'] = np.zeros(data_test_avg.shape[0])
print data_test_avg.shape[0]
data_test_avg.head(10)

for i in range(data_test_avg.shape[0]):
    if i % 1000 == 0: print i
    data_test_avg.loc[i,'avg_l_casual'] = groupby_year_month_weekday_periods.xs(int(data_test_avg.loc[i].year), level ='year'
                                     ).xs(int(data_test_avg.loc[i].month), level ='month'
                                         ).xs(int(data_test_avg.loc[i].weekday), level ='weekday'
                                             ).loc[int(data_test_avg.loc[i].periods)].l_casual

for i in range(data_test_avg.shape[0]):
    if i % 1000 == 0: print i
    data_test_avg.loc[i,'avg_l_registered'] = groupby_year_month_weekday_periods.xs(int(data_test_avg.loc[i].year), 
                                                                               level ='year'
                                     ).xs(int(data_test_avg.loc[i].month), level ='month'
                                         ).xs(int(data_test_avg.loc[i].weekday), level ='weekday'
                                             ).loc[int(data_test_avg.loc[i].periods)].l_registered

for i in range(data_test_avg.shape[0]):
    if i % 1000 == 0: print i
    data_test_avg.loc[i,'avg_l_tot'] = groupby_year_month_weekday_periods.xs(int(data_test_avg.loc[i].year), level ='year'
                                     ).xs(int(data_test_avg.loc[i].month), level ='month'
                                         ).xs(int(data_test_avg.loc[i].weekday), level ='weekday'
                                             ).loc[int(data_test_avg.loc[i].periods)].l_tot

data_test_avg.head(10)

# checking the last three. Intetestingly, the below table has no value for hour == 3  

data_train.loc[115:124]


display(data_test_avg[num_var+cat_var+avg_l_var].head())

data_test_avg[num_var+cat_var+avg_l_var].to_csv('data/test_prep_orig_l_avg.csv', index=False)
